mcp-servers/github-mcp/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
PyGithub==2.5.0
google-cloud-secret-manager==2.20.0
